+++ test/resampling/resampling_schemes/test_BinnedMeanResampling.jl
@@ -1,0 +1,6 @@
+n_draws = 100
+grid = 0:10:500
+resampling = BinnedMeanResampling(grid, n_draws)
+
+@test resampling isa BinnedMeanResampling
+@test typeof(resampling) <: AbstractBinnedSummarisedResampling